# backend/cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "bdUI REST API",
        "contact": {
            "name": "Alex \"bustard\" Provor",
            "url": "https://breezy.su",
            "email": "admin@breezy.su"
        },
        "version": "--dev--"
    },
    "host": "31.56.205.210:8080",
    "basePath": "/",
    "paths": {
...
        "/api/screen/all": {
            "get": {
                "description": "Returns all screen documents.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screens"
                ],
                "summary": "List screens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/echo.SWGScreen"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
        "/api/screen/create": {
            "post": {
                "description": "Creates a new screen document; server generates UUID and returns it as `id`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screens"
                ],
                "summary": "Create screen",
                "parameters": [
                    {
                        "description": "Screen payload",
                        "name": "screen",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echo.SWGScreen"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGCreatedID"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
        "/api/screen/delete": {
            "delete": {
                "description": "Deletes a screen document by its UUID (`id`).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screens"
                ],
                "summary": "Delete screen by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Screen ID (UUID)",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
        "/api/screen/get": {
            "get": {
                "description": "Returns a single screen document by UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screens"
                ],
                "summary": "Get screen by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Screen ID (UUID)",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGScreen"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
        "/api/screen/rewrite": {
            "put": {
                "description": "Updates fields of a screen document by UUID using `$set` semantics.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screens"
                ],
                "summary": "Update (rewrite) screen by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Screen ID (UUID)",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Fields to set",
                        "name": "screen",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echo.SWGScreen"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
        "/api/screen/update": {
            "patch": {
                "description": "Partially updates fields of a screen document by UUID using `$set`; `_id` is preserved.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screens"
                ],
                "summary": "Partially update screen by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Screen ID (UUID)",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Fields to set",
                        "name": "screen",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echo.SWGScreen"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
        "/api/stats/all": {
            "get": {
                "description": "Returns all statistics maps: screen receiving, element clicks, screen clicks.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get aggregated statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGStats"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
...
        "/api/stats/click/screen": {
            "post": {
                "description": "Sets (upserts) click count for screen ID. Intended for client-side aggregated writes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Set screen click count",
                "parameters": [
                    {
                        "description": "Screen click payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/echo.SWGSetClickScreenReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.SWGError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "echo.SWGCreatedID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "echo.SWGError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "description of error"
                }
            }
        },
        "echo.SWGScreen": {
            "type": "object",
            "additionalProperties": true
        },
...
        "echo.SWGSetClickScreenReq": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 12
                },
                "screenId": {
                    "type": "string",
                    "example": "3b2c6f55-1e9e-44e7-8a7e-7f3c5d8a1b23"
                }
            }
        },
        "echo.SWGStats": {
            "type": "object",
            "properties": {
                "clickElements": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "clickScreens": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "screenReceiving": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}